+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.203 ] - 2026/07/04
+
+  * Updated Voice Assist intent imports to delay loading until after initial startup.  This was causing `Detected blocking call to import_module with args ('custom_components.spotifyplus.intent')` exceptions starting with HA 2027.07.0 release.
+
 ###### [ 1.0.202 ] - 2026/06/20
 
   * Added support for new Spotify OAuth refresh token limitations.  The new logic will check for Spotify refresh token `invalid_grant errors`, as Spotify will deactivate refresh tokens after a 6 month period (starting 2026/07/20).  If found, it will place the integration into a "Reauthenticate" state, as well as create a new issue / persistent notification on how to re-authenticate the Spotify refresh token.